Encapsulate global state in TelloApp
